cir_app/requirements.txt
accelerate==1.7.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
comet_ml==3.49.10
configobj==5.0.9
contourpy==1.3.2
cycler==0.12.1
dash==2.16.1
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-holoniq-wordcloud==0.0.4
dash-html-components==2.0.0
dash-image-gallery==0.0.1
dash-table==5.0.0
dash_ag_grid==31.3.1
diffusers==0.33.1
distro==1.9.0
dulwich==0.22.8
everett==3.1.0
filelock==3.18.0
Flask==3.0.3
fonttools==4.58.1
fsspec==2025.5.1
ftfy==6.3.1
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.84.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==5.20.0
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
python-box==6.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
sentry-sdk==2.29.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.0.6
wget==3.2
wordcloud==1.9.4
wrapt==1.17.2
wurlitzer==3.1.1
zipp==3.22.0